dm: core: Use uintptr_t for data in struct udevice_id
Use uintptr_t for data in struct udevice_id as it may use to store pointer
on architecture like CHERI which pointer is not unsigned long integer.

+++ drivers/misc/stm32_rcc.c
@@ -84,14 +84,14 @@ static int stm32_rcc_bind(struct udevice *dev)
 }
 
 static const struct udevice_id stm32_rcc_ids[] = {
-	{.compatible = "st,stm32f42xx-rcc", .data = (ulong)&stm32_rcc_f42x },
-	{.compatible = "st,stm32f469-rcc", .data = (ulong)&stm32_rcc_f469 },
-	{.compatible = "st,stm32f746-rcc", .data = (ulong)&stm32_rcc_f7 },
-	{.compatible = "st,stm32h743-rcc", .data = (ulong)&stm32_rcc_h7 },
-	{.compatible = "st,stm32mp1-rcc", .data = (ulong)&stm32_rcc_mp15 },
-	{.compatible = "st,stm32mp1-rcc-secure", .data = (ulong)&stm32_rcc_mp15 },
-	{.compatible = "st,stm32mp13-rcc", .data = (ulong)&stm32_rcc_mp13 },
-	{.compatible = "st,stm32mp25-rcc", .data = (ulong)&stm32_rcc_mp25 },
+	{.compatible = "st,stm32f42xx-rcc", .data = (uintptr_t)&stm32_rcc_f42x },
+	{.compatible = "st,stm32f469-rcc", .data = (uintptr_t)&stm32_rcc_f469 },
+	{.compatible = "st,stm32f746-rcc", .data = (uintptr_t)&stm32_rcc_f7 },
+	{.compatible = "st,stm32h743-rcc", .data = (uintptr_t)&stm32_rcc_h7 },
+	{.compatible = "st,stm32mp1-rcc", .data = (uintptr_t)&stm32_rcc_mp15 },
+	{.compatible = "st,stm32mp1-rcc-secure", .data = (uintptr_t)&stm32_rcc_mp15 },
+	{.compatible = "st,stm32mp13-rcc", .data = (uintptr_t)&stm32_rcc_mp13 },
+	{.compatible = "st,stm32mp25-rcc", .data = (uintptr_t)&stm32_rcc_mp25 },
 	{ }
 };
