+++ drivers/pci/pcie_fsl.c
@@ -661,16 +661,16 @@ static struct fsl_pcie_data t2080_data = {
 };
 
 static const struct udevice_id fsl_pcie_ids[] = {
-	{ .compatible = "fsl,mpc8548-pcie", .data = (ulong)&p1_p2_data },
-	{ .compatible = "fsl,pcie-p1_p2", .data = (ulong)&p1_p2_data },
-	{ .compatible = "fsl,pcie-p2041", .data = (ulong)&p2041_data },
-	{ .compatible = "fsl,pcie-p3041", .data = (ulong)&p2041_data },
-	{ .compatible = "fsl,pcie-p4080", .data = (ulong)&p2041_data },
-	{ .compatible = "fsl,pcie-p5040", .data = (ulong)&p2041_data },
-	{ .compatible = "fsl,pcie-t102x", .data = (ulong)&t2080_data },
-	{ .compatible = "fsl,pcie-t104x", .data = (ulong)&t2080_data },
-	{ .compatible = "fsl,pcie-t2080", .data = (ulong)&t2080_data },
-	{ .compatible = "fsl,pcie-t4240", .data = (ulong)&t2080_data },
+	{ .compatible = "fsl,mpc8548-pcie", .data = (uintptr_t)&p1_p2_data },
+	{ .compatible = "fsl,pcie-p1_p2", .data = (uintptr_t)&p1_p2_data },
+	{ .compatible = "fsl,pcie-p2041", .data = (uintptr_t)&p2041_data },
+	{ .compatible = "fsl,pcie-p3041", .data = (uintptr_t)&p2041_data },
+	{ .compatible = "fsl,pcie-p4080", .data = (uintptr_t)&p2041_data },
+	{ .compatible = "fsl,pcie-p5040", .data = (uintptr_t)&p2041_data },
+	{ .compatible = "fsl,pcie-t102x", .data = (uintptr_t)&t2080_data },
+	{ .compatible = "fsl,pcie-t104x", .data = (uintptr_t)&t2080_data },
+	{ .compatible = "fsl,pcie-t2080", .data = (uintptr_t)&t2080_data },
+	{ .compatible = "fsl,pcie-t4240", .data = (uintptr_t)&t2080_data },
 	{ }
 };
