+++ drivers/pinctrl/nxp/pinctrl-imx8ulp.c
@@ -20,8 +20,8 @@ static struct imx_pinctrl_soc_info imx8ulp_pinctrl_soc_info1 = {
 };
 
 static const struct udevice_id imx8ulp_pinctrl_match[] = {
-	{ .compatible = "fsl,imx8ulp-iomuxc0", .data = (ulong)&imx8ulp_pinctrl_soc_info0 },
-	{ .compatible = "fsl,imx8ulp-iomuxc1", .data = (ulong)&imx8ulp_pinctrl_soc_info1 },
+	{ .compatible = "fsl,imx8ulp-iomuxc0", .data = (uintptr_t)&imx8ulp_pinctrl_soc_info0 },
+	{ .compatible = "fsl,imx8ulp-iomuxc1", .data = (uintptr_t)&imx8ulp_pinctrl_soc_info1 },
 	{ /* sentinel */ }
 };
