+++ drivers/pinctrl/nxp/pinctrl-vf610.c
@@ -16,7 +16,7 @@ static struct imx_pinctrl_soc_info vf610_pinctrl_soc_info = {
 
 static const struct udevice_id vf610_pinctrl_match[] = {
 	{ .compatible = "fsl,vf610-iomuxc",
-	  .data = (ulong)&vf610_pinctrl_soc_info },
+	  .data = (uintptr_t)&vf610_pinctrl_soc_info },
 	{ /* sentinel */ }
 };
